Added static_cast to narrow conversion

+++ helpers.cpp
@@ -39,9 +39,9 @@ void edges(int height, int width, RGBTRIPLE **image) {
 void sepia(int height, int width, RGBTRIPLE **image) {
     for (int i {0}; i < height; ++i) {
         for (int j {0}; j < width; ++j) {
-            int sepiaRed   {(.393 * image[i][j].rgbtRed) + (.769 * image[i][j].rgbtGreen) + (.189 * image[i][j].rgbtBlue)};
-            int sepiaGreen {(.349 * image[i][j].rgbtRed) + (.686 * image[i][j].rgbtGreen) + (.168 * image[i][j].rgbtBlue)};
-            int sepiaBlue  {(.272 * image[i][j].rgbtRed) + (.534 * image[i][j].rgbtGreen) + (.131 * image[i][j].rgbtBlue)};
+            int sepiaRed   {static_cast<int>(.393 * image[i][j].rgbtRed + .769 * image[i][j].rgbtGreen + .189 * image[i][j].rgbtBlue)};
+            int sepiaGreen {static_cast<int>(.349 * image[i][j].rgbtRed + .686 * image[i][j].rgbtGreen + .168 * image[i][j].rgbtBlue)};
+            int sepiaBlue  {static_cast<int>(.272 * image[i][j].rgbtRed + .534 * image[i][j].rgbtGreen + .131 * image[i][j].rgbtBlue)};
 
             image[i][j].rgbtRed   = sepiaRed > 255 ? 255 : round(sepiaRed);
             image[i][j].rgbtGreen = sepiaGreen > 255 ? 255 : round(sepiaGreen);
